pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = {}

[project]
name = "environ"
version = "0.0.1" 
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "requests",
    "matplotlib",
    "seaborn",
    "openai",
    "beautifulsoup4",
    "lxml",
    "python-dateutil",
    "tqdm",
    "python-dotenv",
    "tiktoken",
    "torch",
    "scikit-learn",
    "neuralforecast",
]

[project.optional-dependencies]
dev = ["pylint"]